+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Chat.ChatActivity" />
         <activity android:name=".Setting.PersonalSettingActivity" />
 
         <provider
